Create player.js
Heads up, JavaScript can't do SQL stuff on its own. Normally, you'd use a library that helps with this, like an ORM or a raw connection library. For example, if you're using Node.js with MySQL, it'd look like this: let result = await runMySqlQuery("SELECT ...");
But you'll need to do the SQL part yourself.
Make sure to put in the right SQL command where it says TODO. I will add to the documentation on how you can do this.